requirements.txt
asgiref==3.6.0
boto3==1.35.7
botocore==1.35.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.1
django-autocomplete-light==3.9.7
django-cassandra-engine==1.8.0
django-smtp-ssl==1.0
django-storages==1.14.4
docutils==0.16
feedparser==6.0.12
fuzzywuzzy==0.18.0
geomet==0.2.1.post1
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psycopg2-binary==2.9.7
pyasn1==0.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
rsa==4.7.2
s3transfer==0.10.2
scikit-learn==1.2.2
scipy==1.10.1
scylla-driver==3.26.2
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.19
